+++ 2D/Perco-Schramm.cpp
@@ -73,7 +73,7 @@ private:
     vector<bool> cols, mask;
     const double omx{sqrt(3.0)};
 
-    cpx thepos(size_t i) { return cpx(omx * (((i / w) % 2) + 2 * (i % w)), 3 * (i / w)); }
+    cpx thepos(size_t i) { return cpx(omx * double(((i / w) % 2) + 2 * (i % w)), 3 * double(size_t(i / w))); }
 
     size_t follow(size_t base, size_t dir) {
         static const vector<int> fola = {1, int(w), int(w) - 1, -1, -int(w) - 1, -int(w)};
